pixman: Add "-Wno-unknown-attributes" and cc_library_headers
1. "-Wno-unknown-attributes" in cflags Pixman fails to build on arm such
as aosp_cheetah

```
In pixman-glyph.c
__attribute__((__force_align_arg_pointer__))
```
This conditional is problematic because it applies the attribute when
NOT on x86_64, which incorrectly includes ARM. Adding the flag will make
the attribute silently ignored on ARM platforms.


2. cc_library_headers In pixman.h on line 72, it's including the file
with angle brackets: Angle brackets tell the compiler to look in system
include paths, not in the local project paths. The header library and
linking against allows the build system to handle the bracket includes

Bug: b/413089179
Test: lunch aosp_cheetah-trunk_staging-userdebug && m pixman
Change-Id: I42a3a9a8756c9f7b368313b8faaea72903f6aa82